updated RHS side with functions
